+++ Client/src/hooks/useKakaoLoginQuery.tsx
@@ -1,5 +1,5 @@
-import { useQuery } from '@tanstack/react-query';
+// import { useQuery } from '@tanstack/react-query';
 
-const useKakaoLoginQuery = () => {
-  return useQuery(['kakaoLogin']);
-};
+// const useKakaoLoginQuery = () => {
+//   return useQuery(['kakaoLogin']);
+// };
